+++ KeeAnywhere/StorageProviders/Dropbox/DropboxHelper.cs
@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Dropbox.Api;
 
 namespace KeeAnywhere.StorageProviders.Dropbox
 {
@@ -19,10 +20,21 @@ namespace KeeAnywhere.StorageProviders.Dropbox
        */
 
         //TODO: Change API keys!!!
-
         internal const string DropboxFullAccessClientId = "dummy";
         internal const string DropboxFullAccessClientSecret = "dummy";
         internal const string DropboxAppFolderOnlyClientId = "dummy";
         internal const string DropboxAppFolderOnlyClientSecret = "dummy";
+
+        public static DropboxClient GetApi(string accessToken)
+        {
+            var config = new DropboxClientConfig
+            {
+                HttpClient = ProxyTools.CreateHttpClient()
+            };
+
+            var api = new DropboxClient(accessToken, config);
+
+            return api;
+        }
     }
 }