+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### New
 - \#10 Access to Amazon Cloud Drive
 - \#11 Access to Box
+- \#16 Proxy support (with auth) feature request
 - \#26 AWS S3 support
 - Possibility to donate